Update transaction_test_helpers.rs
so 10k for MAX_GAS_AMOUNT ok but is this thing not throwing an error leaving that underscore in it?
i recommend getting rid of this and working only with a clean number.

but i also wonder where this gas price 1 relates to? the only thing that makes sense for downlink compatiblitiy in old monetary thinking would be using 1 penny or cent contracting it there to. its the lowest point for the coins and gas is the smallest units for paying for transactions.