[Consensus] Separated util module
This diff does not change any logic, it simply moves mutex_map and time_service to new util directory.
Motivation for this change is that we probably going to have more 'utility' modules in consensus, so it's good to not put them in top level directory, but instead organize into separate dir.