Move to newtype wrapper for all units relating to gas
This implements units around gas using newtype wrappers that hold their underlying type.

This then allows us to statically enforce that operations between these newtype-wrapped units are performed in a principled manner using the `GasAlgebra` trait -- e.g. if `AbstractMemorySize`'s underlying type was changed to a `u32` things would break at complile-time when we tried using this in conjunction with a `GasUnit` that uses a `u64` to hold its value.