[terraform] Support running stable in prod account
This supports a new `prod` workspace which will run in the new
account. TF state will be stored in the same S3 bucket, but references
to the global remote state are to the prod S3 bucket. Working in the
correct AWS account is enforced with `allowed_account_ids`. The Route53
domain is moved back to a reference to global state.