[Storage] persist LedgerCounters on transaction block commition
ChangeSet and SealedChangeSet are introduced to keep track of counter bumps across multiple store API calls and make sure the bumps are converted to db writes before committing.

With this we now know exactly how many nodes we have in the state merkle tree.
After blob retire counter is added and maintained, we can know exactly how many accounts we have in the system.