[dev-infra] introduce anchor and x
Introduce `anchor`, the beginnings of a toolchain meant to make the
lives of developers on the Libra project a little bit easier and `x` a
light-weight wrapper around `anchor`.

Why `anchor`? Anchors are attached to a "chain" on ships to prevent them
from drifting. Ships sail the ocean and the symbol for Libra is meant to
look like water on the ocean.