[language] Fix type sig definitions for Eq and Neq
Before Todd's recent change to equality we were using Rust's version of
equality, which, in particular meant that it was "legal" to pass in e.g.
`1 == true` to the VM. However, with the recent update this is no longer
the case so we need to update the typed call details for the bytecode
instructions to encode this new restriction.