[consensus] ProcessProposalResult now contains a better signal for success
When a validator receives a proposal, it is processed and a ProcessProposalResult enum is returned as a result.
This enum either indicates that there is more work to be done to process the block, or that there is nothing to be done (the proposal was processed).
In the latter case, the enum does not distinguish a failure to process the proposal versus a proposal fully validated and voted on.

The reason is that even though the proposal is then called winning_proposal, we still cannot fully know if we will eventually vote on the proposal as there is more logic that could still prevent the validator to vote on the proposal.

Yet, this PR adds a boolean inside the enum, with true indicating a proposal that was processed up until it was decided to be sent to the proposer_election component, and false indicating otherwise.

This change is useful for fuzzing (we need to know if we can call further functions on the proposal or if some of the validation has failed) and for further testing.