[Consensus] Introduce SyncInfo struct for accumulating sync information
Summary:
This commit doesn't change anything in the logic: it's the first preparation for the changes required for being able to send the missing pieces of information upon receiving of outdated timeout messages.
Today sync info is scattered across several fields both in the `TimeoutMsg` and in `Proposal`. This commit introduces a new struct `SyncInfo`, which is going to be included in both `TimeoutMsg`, `Proposal`, and could be sent as a standalone message.

Test: existing unit test coverage

Issue: #311