Add checks for exhaustive coverage of instructions
Currently there are no checks that ensure that every bytecode instruction is included in the CostTable.
This diff adds a check that will be run in debug builds to ensure that exactly 53 instructions (without duplicates) are present in the memory_table and compute_table.
This check is run during invocation of cargo test and thus can prevent a situation where a new instruction is added to the bytecode but is not added to the CostTable.