[Proof] Compute accumulator root hash using frozen subtrees
This functionality will be used when we need to verify the proof
showing that two accumulators are consistent -- one of them can be
obtained by appending certain items to the other.

Given that the scratchpad code also uses this functionality, we first
extract the code. We also implement it differently without using the
`Position` library.