Continuous Fuzzing Integration with Fuzzit
This commit inroduce continuous fuzzing workflow with
the following additions:

1) on every new code pushed to master the fuzzers are built and
sent to Fuzzit servers where they are continuously run and alert
the relevant people if a new bug is found.
2) on every pull-request a quick regression test is run using
the corpus which is generated on Fuzzit. This will fail the CI
and prevent new or old bugs from being merged.

Signed-off-by: Yevgeny Pats <yp@fuzzit.dev>