[language] module for block metadata
*Motivation*

First step toward representing block metadata (height, time, signers, etc.) in the global state. This commit adds a new `Block` module with a `Block.T` resource that (for now) contains only the height. The module and a `Block.T` resource at the Association address are included in the genesis blob.

The `Block` module contains a `prologue` procedure for setting the block metadata. In the future, a special transaction invoking this procedure will be the first tx executed in every block.

*Test plan*

New E2E tests for reading the block height and invoking the block prologue.