[crypto] Enable derivation of cryptographic traits on any enum of cryptographic keys
*What this changes*:
The new crypto API is based on static type safety by default. Nonetheless, it's sometimes necessary to accept several signing / verification schemes at runtime.
For this we can define an enum to list the alternatives for each of (Public key, Pricate key, Signature), and write the cryptographic trait implementations that trivially dispatch to each variant. This is demonstrated in crypto::unit_tests::cross_test (and used in production in secret_service::crypto_wrappers).
This PR makes this boilerplate a derive macro, which makes any such set of enums definable in not much more than the space it takes to define the enum.
This also converts secret_service::crypto_wrappers to these derive macros.

*Why this is better*:
- developers can define a new enum to locally admit whichever combination of signing schemes they think is better, if needed,
- the dispatch is automated and thus less tedious and error-prone,

*Why this is worse*:
- strong dependency on dtolnay's syn crate

*Testing*:
- converted `secret_service::crypto_wrappers` for structural compliance