[Executor] Handle empty chunks during state synchronization
Currently, if a transaction chunk is not the last chunk corresponding to
the ledger info, and it is empty (or it is not empty but all the
transactions in the chunk have been committed, for example this could
happen if a chunk is sent to executor twice), storage would return an
error since it expects either new transaction or new ledger info.

Fix this by not calling `save_transactions` in this case.