[ci][configs] Fail commit workflow if the code causes genesis.blob without modifying it
We store configs in our repository, but do not have a process to keep them
in sync with code. This commit adds a check in CI to fail commits that
cause this inconsistency.

This is *limited* because our config generation is not stable, so we'll start
with checking only `genesis.blob`.