[types] Make ValidatorVerifier a trait
MOTIVATION:

For fuzzing, it would be nice if I could create a fake ValidatorVerifier that verifies for any signature.
This would allow me not to care about the validity of a signature (which is not what I'm testing).

This PR creates a `ConsensusVerifier` trait, that is implemented by ValidatorVerifier.
It also defines a `FakeVerifier` struct that implements the trait. In future PRs it will be used for fuzzing and tests.

TEST PLAN:

```
cargo test --lib
```