[language][not for commit] VM benchmark
This is a way to run benchmark against the VM.
I am not going to commit this as a better way is once we modularize the VM
enough to make it really easy to run the interpreter without spinning a
transaction. For now though this is good enough for someone that wants
to try some benchmark.
This probably comes with some friction as it may not build as things change,
and other possible problems, so it's just as a possible refrence...