[types] Deprecate Transaction::Program
This diff is going to use our new model of Transaction::Script and Transaction::Module to do simple payment transactions in the libra codebase. This diff will NOT be a breaking change though as we still keep the functionality of processing Transaction::Program in the VM so that we will have some time to break the backward incompatability there.