[language] Move the interpreter code out of transaction processor.
This is the first step in a series of cleanup to make the VM more controllable from the outside
and more isolated from the transaction flow and the logic of transactions.
In the process simplified things a bit particularly with the logic of call frames.
For susequent cleanups it's important to remove the Program transaction type and remove the opcodes
related to account creation (CreateAccount) and transaction metadata
(eg: GetTxnSenderAddress, GetTxnSequenceNumber, etc.).

Code for gas metering and cost synthesis has been factored out into a couple of "special" impl;
that should help to see what is needed once that code is pulled into the create or closer to the VM.

I removed entirely a unit test that was relying too much on implementation details. The interpreter
is heavily tested already but once we clean things up more and we have a good way to initialize
an interpreter we will add tests for specific code sequences and instructions. It did not seem that
vital or helpful to have those tests right now.

I changed a bit the native dispatching and plan to make more changes to make it easier to
have error code flow out of native functions.

I changed the From trait for conversions to return a VMResult<> so they can make the error
flow naturally. I think it would be better to call them Cast<> and have an internal Cast trait
instead of From and I will do it in a different diff.

This also logs a dump of the interpreter state on invariant errors which gives good info about
what is going on.