[Reconfiguration] Emit a ValidatorSet.Change event when ValidatorSet.reconfigure() is called
Eventually, the Move ValidatorSet module will trigger reconfiguration by emitting an event each time the on-chain validator set changes. The executor will know to look for this event and pass it on to consensus to actually do the reconfiguration.

This PR takes the first step by implementing the Move logic to emit an event on validator set change. Checking for the event in the executor and forwarding it to consensus will follow.

Additionally, I had to fix two bugs that were revealed by the tests I added:
- The canonical serializer created EventKey's whose byte representation was different from Move EventKey's (in the EventKey.new_from_address procedure). I fixed this by mimicking the Move logic in Rust for now because it was less disruptive, but we should strongly consider making the Move ByteArray.concat logic (the source of the problem) match the behavior of the canonical serializer in the future.
- The serialization of the ValidatorPublicKeys struct in Rust/Move did not match because of different field orderings. I made the field orderings consistent to solve the problem.