[language] Add counters to track execution time and gas usage stats
Add a number of different timers in order to track how much time is
being spent in very parts of the transaction flow. In particular, in:
* The bytecode verifier
* Transaction validation (aka prologue)
* Transaction execution
* Transaction epilogue

We also add histograms to track gas for these sections as well:
* Total gas usage for the transaction
* Gas usage for transaction execution

The primary purpose for these statistics is to start collecting data to
measure how "good" the gas metering is -- in particular correlation
between the gas used and the time spent processing the transaction.