[libra-swarm] dump logs on failure
Teach LibraSwarm to dump logs for each validator in the swarm when
panicking and `LIBRA_DUMP_LOGS` environment variable is set. This
enables us to gain more insight into test failures in CircleCi since we
don't have any way to grab the logs of a failed run off the build
machines.

Closes: #1049