[Common] Remove grpcio-client crate
First, there's currently no code that uses this.

Second, the original intention of this was that we create a trait for
each GRPC client, so in tests we can do mock clients by implementing the
trait. However, given that we should use a Rust wrapper for the GRPC
clients, any use case should manually create a trait like StorageRead
https://github.com/libra/libra/blob/096245b0343231544ae10f1cc9b7fd15d396afff/storage/storage_client/src/lib.rs#L297
instead of using codegen.