[consensus] Separate out safety rules into its own crate
The intent is to build a TCB within side this space. Had to add a
wrapper to update counters. The exact construction of this will be more
intuitive once we actually get to the point where the interface is
stable, such as returning an optional value if any state has changed.

Tests have been rewritten to not depend on consensus.
It was discussed to delete the proptest. We'll add back something that
maintains the intent of that test but without the reliance on blockstore
and is a bit more intuitive.