[language] new lalrpop-replacement parser
This is a hand-written recursive descent parser to replace the
lalrpop parser that we have been using so far. The lexer and
error handling need some serious work, but for now, I'm using
code that is similar to what lalrpop generated. Both of these
should be completely replaced. I tried pretty hard to make this
a drop-in replacement for the current parser, so it maintains
the same structure and behavior as the old lalrpop file.