[consensus] Signatures encoded into consensus blockchain
Currently, when serializing blocks (used for hashing), only the block
id of the QuorumCertificate is hashed, but not the signatures of the
QuorumCertificate.  This means that the consensus blockchain can
disagree on the signatures validating the consensus blockchain between
validators.  This also makes it hard to assign rewards to validators
that participate in consensus since there isn't agreement on the
signatures.

This PR adds support for encoding all the QuorumCertificate
information into the canonical hashing method (serialize()), so that
the consensus blockchain agrees on the QuorumCertificates and Blocks
that have been committed.  There is also a test to ensure this
behavior and verified it was failed before this change and now passes.
LedgerInfoWithSignatures now has a BTreeMap of signatures to make it
easier to do canonical serialization.