[Consensus][tiny] Fix a chained_bft_smr_test
Summary:
basic_commit_and_restart test got flaky after https://github.com/libra/libra/pull/1245: we're counting the vote messages in the test, hence the vote message corresponding to round 20 may or may not update the root of the tree to round 18 as observed by the test (it is racing with the test check). We can only assert that it's at least 17.

Testing:
the tests pass