[lcs] introduce a Serde implementation of LCS
Currently we have LCS implemented as a custom framework using a set of
traits `CanonicalSerialize` and `CanonicalDeserialize` that must be hand
implemented for each type that wants to use the LCS format. This is very
error prone as its easy to mess up the implementation as well as
difficult when we need to refactor a type because we also need to
rewrite the ser/de implementations to match the change.

This patch introduces a Serde `Serializer` and `Deserializer`
implementation for the LCS data format. This will enable us, in a follow
up patch, to stop hand-writing ser/de implementations and instead rely
on being able to derive `Serialize` and `Deserialize` for most types.