[Consensus] Introduce highest timeout certificate into block store
Summary:
Note that it's an intermediate step on the way to using TimeoutCertificates for pacemaker.
In this change we are adding the highest timeout certificate to the block tree. This is going
to be the timeout certificate, which is persisted in the ConsensusDB instead of the pacemaker timeout certificate.
The highest timeout certificate is still not used by any participant, to be added in the followup change.

Testing:
Current unit tests.

We are introducing the highest timeout certificate into the block store an

Closes: #1302
Approved by: zekun000