[Proof] Make verify a method of AccumulatorProof
Right now there are too many `verify_xxx` functions in `proof/mod.rs`. I
think making a `verify` method would be nicer. Some existing structs
like `EventWithProof` already has it.

Note that we are making the hasher as a generic parameter of the proof,
instead of having a generic parameter on the function.

Closes: #1286
Approved by: msmouse