[language][execution] Bubble up VM startup errors
Certain conditions may cause the VM to be unable to start. This commit
allows the VM to signify this back up to the executor, and from there
for it to bubble that back out to consensus. After this, consensus can
determine how it wants to handle this case.

Closes: #1381
Approved by: lightmark