[CI] code coverage by grcov
We need to track the code coverage to help us ensure the basic testing
is in place to exercise all the code.

We are adding grcov to CI pipeline to collect and aggregate code
coverage information. The coverage result is uploaded to codecov.io,
which can display the aggregate and integrate the result into code
review process.

Read more about grcov at https://github.com/mozilla/grcov