fixup! [language] new lalrpop-replacement parser
Clean up the lexer: get rid of the double underscore symbol
prefixes, add functions to access the current token info, and rename
from "matcher" to "lexer".

This also fixes a problem with the end location for Spanned content:
the previous implementation (and possibly the lalrpop code, although
I'm not sure) was setting the end location as the start of the
following token, so that the intervening whitespace was included in
the span. The tokens already included the end offset, but there
wasn't a good way to access that info with the way that the parser
advances past tokens as they are used. This change adds a new
"previous_end_loc" function to get at the end location for the
previous token.

As suggested in the code review, this change also adds a "spanned"
function to the parser to simplify construction of Spanned structures.