[language] native functions NativeReturnStatus change
The comment in `language/vm/vm-runtime/vm-runtime-types/src/native_functions/dispatch.rs`
for `NativeResult` should describe this change well. Looking at native functions implementations
should also provide a clear view of the change.
Effectively instead of returning a `NativeReturnStatus` from native functions we return a
`VMResult<NativeResult>` which plugs into the way we report errors a bit more naturally.
A `NativeResult` carries a cost always and it's the way errors should be returned if they
are intended to charge the user. Typically for invariant violation a `VMResult` should be returned.
There is more code deleted than added so it seems a change in the right directions.
Benchmarks seem to show a minor win (less than 5%) - which may not be a win at all - but
at least it does not look like this regresses anything. I am running more tests on
a devserver to get a better sense.

I am curious if people think this is not a good direction to go.
There are reasons the model previously always returned a `NativeReturnStatus`
however I don't think this lose clarity and, in fact, it should make things
more natural as I mentioned.
But others may have a different perspective...