[Consensus] Highest ledger info processing using the new state sync flow
Summary:
In the new state sync flow we're first trying to state sync to the given target or higher,
then retrieve the ancestors of the new root and rebuild the tree.
In this change we are not using the flow yet but introduce the first pieces.
There is no proper testing and there is no proper persistency.

Testing:
The new feature is not in use yet, tests to be added at the level of SMR.