[CI] add conditional docker build to commit verify
When we make changes to these docker files, we want to verify them in
the commit work flow.  Or we end up breaking nightly. This is a
follow-up to 243f535.

Here we added a conditional docker build job to the commit verify work
flow.  The docker build job will kick off iff the PR contains change in
any file matches `*.Dockerfile`.  Once triggered, the work flow will
build each of the updated docker files.