[language] remove CreateAccount bytecode
Remove the bytecode CreateAccount in favor of a native function.
This is mostly deleting code and a decent change in our proptest for e2e tests.
We have now enabled gas on create account. That was problematic because create account
has different gas cost depending on whether the event counter in the account that
sends the transaction had been created or not. So we need to track that operation.

This specific PR deprecates CreateAccount, in other words it fails CreateAccount
but leaves existing scripts identical. The idea here is that everything that does
not use CreateAccount keeps working fine.