[Consensus] Block retrieval of descendants of a committed id
Summary:
This diff introduces a new type of block retrieval: retrieval of the descendants of a committed id.
The idea is to return the chain that starts with the highest round, end at a given target id,
and contains a ledger info that proves a commit of a given target id.

We thus introduce two types of BlockRetrievalMode (Ancestors and Descendants).
Most of the business logic is in the BlockTree, which is running DFS in order to find the chain that would fit the request.

Tests:
Added unit tests for the retrieval of the descendants of the committed id both at the leve of BlockTree and EventProcessor.