[x] skip running cargo if we have no packages to run
Currently if an empty iterator is passed to `run_on_packages_together`,
we'll happily run the cargo command with no `--package` args. This patch
fixes this behavior and instead checks to see if we have any package
args and does an early return if the provided iterator is empty.

Closes: #1578
Approved by: metajack