[lcs] introduce from_bytes_seed
Introduce the `from_bytes_seed` method which enables stateful
deserialization from a `&[u8]` based on some initial state `seed`. This
is the more idiomatic and recommended way to perform stateful
deserialization using serde.

Now that there is a way to do stateful deserialization, remove the
public export of the Deserializer type defined in LCS.