[Storage] Separate get_child logic from get_siblings
Currently get_child_with_siblings does two things given an input `n`:

1. Returns the child's node key.
2. Returns the necessary siblings.

Most of the logic is about getting the siblings, which is much more
complex than getting the child.

Rename this function to `get_siblings` and move the logic to get the
child outside. So the tests are simpler and it's possible to reuse this
code somewhere else.