LedgerInfo commit information aggregated in a BlockInfo struct
Summary:
The fields of LedgerInfo that describe the committed status of the Ledger are in fact
identical to the fields of the block metadata that Consensus is carrying around (the only
field of BlockInfo that is currently not present in LedgerInfo is a round).
Any update to the LedgerInfo would have to be mirrored in BlockInfo because TCB needs to verify & sign it.
Hence, this change is aggregating the LedgerInfo fields in the BlockInfo.
We had to move BlockInfo from consensus types to libra types as a result of that.

Testing: this is supposed to be a noop, existing unit test coverage

Ref #1604