[consensus] Move toward PersistentStorage interface for SafetyRules
ConsensusState isn't really a store, SafetyRules needs a store. So this
replaces the code to leverage a config backed storage unit and all the
code mods that go with it.

Note: because of the fact that consensus is multithreaded,
PersistentStore must be both Send + Sync

This commit also refactors the code within safety rules to take on a
more library style approach as there are more features now within the
code.