[mempool] commit old transactions on transaction insertion
Fixes #1625
When node does state sync, transactions that are committed through this sync are not removed from mempool.
This means that even though transactions are committed, client can not use this node to submit new transactions, until they expire, which can take significant amount of time.

We do state sync quite often, so this reproduce routinely during cluster test - we had to add retry on transaction submission because of this problem