[State sync] State synchronization with SynchronizerState
Summary:
Our understanding of StateSynchronizer requirements and storage behavior has grown organically: this diff is an attempt to update the code accordingly.
The main changes in this commit are as follows:
1. Introduce a `SynchronizerState` struct, which distinguishes between the latest committed transaction and the latest LedgerInfo.
While remote peers can be helped with information covered by the LI, chunk requests should take into account all the committed transactions.
2. Choose an approach of active re-syncing of the local state with the storage, as an alternative of managing the latest commit in memory.
The main advantage of bookkeeping the latest commit index is the potential performance gains, however, the re-syncing approach is simpler to reason about and much less error-prone,
especially if there are concurrent modifications to the storage.
Caching of storage data can be re-introduced again if we discover that there are performance penalties of re-syncing once per chunk (my gut feeling is that with the reasonable chunk sizes this cost is negligible).
3. The role of peer manager for penalizing validators should be adjusted in the followup diff. Today's approach used to penalize validators, which sent
chunks that a node believe it never asked for, however, the bookkeeping would not work correctly for chunk retries or empty chunk responses.
This change removes some of the penalizing code and adds corresponding TODOs.
4. General cleanup

Testing:
the current tests are passing, but I'd love to add a comprehensive set of unit tests in a followup diff.

Issues:
ref #899