[State sync] PeerManager aware of retries
Summary:
The peer manager bookkeeping didn't take into account the fact that the same version could retry, which could cause to:
* unintended penalty of peers with late responses
* counter for the chunk processing duration that takes into account the last attempt only

In this change we keep the pending requests per peer in the `PeerInfo` for penalizing the wrong responses.
We also keep both the first and the last request timestamp in a `ChunkRequestInfo` for calculating the request processing duration.

Testing:
Added unit tests for PeerManager testing the new functionality

Issues:
ref #899