[config] Update terraform configs
It is really challenging to produce the exact same random keys in the
new config generation framework. As a result, we get this massive commit
to reset the keys. Also as I was reseting the keys I stumbled upon two
improvements:
- Specify where to save the consensus key while generating validator
nodes as there is an interesting dependency between the peer id and the
consensus key due to how our terraform config generation works
- Do not save default consensus keys otherwise fullnode keys will be
saved to disk