[config] split prepare and load
When first working on the configs, prepare and load made sense. However,
there are cases where a config may be fully populated having a data_dir
set. This would then result in a very confusing crash if the dir was
set. Furthermore, setting the dir would overwrite the contents of the
loaded config, which is a strange side effect.