[consensus][reconfig] ensure deterministic timestamp for genesis
We may have arbitrary long suffix blocks after reconfiguration and
many of them could commit the reconfiguration block and derive the
genesis.
Thus to guarantee the deterministic genesis, the timestamp of all of
those commit have to be the same.

This change we ensure all suffix blocks have the same timestamp as
the reconfiguration block. Also all nil blocks have the same timestamp as parent.

Changing the executor to use epoch/round to compare instead of
timestamps.