[consensus][reconfig] notify epoch change upon sync
Normally a node aggregates the end-epoch QC and broadcasts the proof
to everyone else.

While there's edge cases that nodes might receive a SyncInfo that has
end-epoch QC and needs to start new epoch too, e.g. sending a older
SyncInfo to the node that's committing the reconfiguration block and
reply a SyncInfo with end-epoch QC.

This change addresses the problem and make sure nodes can switch epoch
after sync to the reconfig block.