Add storage API to retrieve items without sending back the epoch change proofs.
Summary:
New transactions in Mempool used to go through a series of vm_validator checks, and each such check used to call update_latest_ledger_info.
The update_latest_ledger_info might work hard to retrieve the epoch change proofs in case the client version in the request is not properly set.
In our case case client version was always set to 0.
The irony is that vm validator does not really use any of these proofs: it's not very critical. In the case the local storage
is corrupt the bad transaction might infiltrate the consensus but would still fail later during the execution.

We now have two options: either mempool managing a state with the known state and sending the right client version, or having a lightweight API
for retrieving the items without the epoch proofs. I added an internal storage API for retrieving the items without the epoch change proofs.

Testing: existing test coverage.

Issues: ref #1996