[language][move] Merge "ReturnAbort" expressions into top-level expressions
The Move grammar had a separate non-terminal for ReturnAbortExp, which was
apparently a workaround for lalrpop's inability to handle the shift/reduce
conflict of the "dangling else" problem. With the current top-down parser,
that ambiguity is no longer significant, so we can simplify the grammar
and avoid the need to put assignments and control expressions inside braces
or parenthesis when they are used as the body of another control expression.