[Storage] Expose the LedgerInfo by version by API
Summary:
We need an API for finding the LedgerInfo corresponding to the given version.
The storage API is going to return the lowest LedgerInfo that is covering the given version. Specifically for the waypoint at the epoch boundary the returned LedgerInfo would match the waypoint version.

Testing: current test coverage

Issues: ref #1384